Move current_lesson_plan* to controller concerns
